requirements.txt
fastapi
uvicorn
google-genai
langchain
langchain-community
chromadb
python-dotenv
sentence-transformers